First Round of Updates
After a struggle getting Zend to play nice with my previous apache, and
figuring out how Netbeans (the IDE I use) works with Zend (which as it
turns out, it works very well with it...), I finally got some of it to
work.

In the original Java EE project, we can SHOW departments, employees, and
projects. Then, we can click on them to VIEW various information related
to them.

So far I got the first part (the SHOW) to work.  I've yet to make the
VIEW files but I 'll probably do them tomorrow.